requirements.txt
asgiref==3.2.8
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-enum-choices==2.1.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
django-storages[azure]==1.9.1
ecdsa==0.15
idna==2.9
jws==0.1.3
lxml==4.5.1
oauthlib==3.1.0
Pillow==7.1.2
pyasn1==0.4.8
python-jose==3.1.0
python3-openid==3.1.0
pytz==2020.1
psycopg2-binary==2.8.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
uvicorn